+++ gmusicapi/clients/webclient.py
@@ -17,16 +17,18 @@ class Webclient(_Base):
     to upload).
 
     Any methods in this class that are duplicated by
-    the :class:`Mobileclient` should be considered deprecated.
+    the :class:`Mobileclient` are deprecated, and will generate a
+    warning at runtime.
+
     The following methods are *not* deprecated:
 
+        * :func:`create_playlist`
         * :func:`get_registered_devices`
         * :func:`get_song_download_info`
         * :func:`get_stream_urls`
         * :func:`get_stream_audio`
         * :func:`report_incorrect_match`
         * :func:`upload_album_art`
-        * :func:`create_playlist`
     """
 
     _session_class = gmusicapi.session.Webclient
@@ -277,11 +279,16 @@ class Webclient(_Base):
 
         return url
 
+    # deprecated methods follow:
+
     @utils.accept_singleton(basestring)
     @utils.enforce_ids_param
     @utils.empty_arg_shortcircuit
+    @utils.deprecated('prefer Mobileclient.delete_songs')
     def delete_songs(self, song_ids):
-        """Deletes songs from the entire library. Returns a list of deleted song ids.
+        """**Deprecated**: prefer :func:`Mobileclient.delete_songs`.
+
+        Deletes songs from the entire library. Returns a list of deleted song ids.
 
         :param song_ids: a list of song ids, or a single song id.
         """
@@ -290,126 +297,15 @@ class Webclient(_Base):
 
         return res['deleteIds']
 
-    def get_all_songs(self, incremental=False):
-        """Returns a list of :ref:`song dictionaries <songdict-format>`.
-
-        :param incremental: if True, return a generator that yields lists
-          of at most 2500 :ref:`song dictionaries <songdict-format>`
-          as they are retrieved from the server. This can be useful for
-          presenting a loading bar to a user.
-        """
-
-        to_return = self._get_all_songs()
-
-        if not incremental:
-            to_return = [song for chunk in to_return for song in chunk]
-
-        return to_return
-
-    def _get_all_songs(self):
-        """Return a generator of song chunks."""
-
-        get_next_chunk = True
-        lib_chunk = {'continuationToken': None}
-
-        while get_next_chunk:
-            lib_chunk = self._make_call(webclient.GetLibrarySongs,
-                                        lib_chunk['continuationToken'])
-
-            yield lib_chunk['playlist']  # list of songs of the chunk
-
-            get_next_chunk = 'continuationToken' in lib_chunk
-
-    @utils.enforce_id_param
-    def get_playlist_songs(self, playlist_id):
-        """Returns a list of :ref:`song dictionaries <songdict-format>`,
-        which include ``playlistEntryId`` keys for the given playlist.
-
-        :param playlist_id: id of the playlist to load.
-
-        This will return ``[]`` if the playlist id does not exist.
-        """
-
-        res = self._make_call(webclient.GetPlaylistSongs, playlist_id)
-        return res['playlist']
-
-    def get_all_playlist_ids(self, auto=True, user=True):
-        """Returns a dictionary that maps playlist types to dictionaries.
-
-        :param auto: create an ``'auto'`` subdictionary entry.
-          Currently, this will just map to ``{}``; support for 'Shared with me' and
-          'Google Play recommends' is on the way (
-          `#102 <https://github.com/simon-weber/Unofficial-Google-Music-API/issues/102>`__).
-
-          Other auto playlists are not stored on the server, but calculated by the client.
-          See `this gist <https://gist.github.com/simon-weber/5007769>`__ for sample code for
-          'Thumbs Up', 'Last Added', and 'Free and Purchased'.
-
-        :param user: create a user ``'user'`` subdictionary entry for user-created playlists.
-          This includes anything that appears on the left side 'Playlists' bar (notably, saved
-          instant mixes).
-
-        User playlist names will be unicode strings.
-
-        Google Music allows multiple user playlists with the same name, so the ``'user'`` dictionary
-        will map onto lists of ids. Here's an example response::
-
-            {
-                'auto':{},
-
-                'user':{
-                    u'Some Song Mix':[
-                        u'14814747-efbf-4500-93a1-53291e7a5919'
-                    ],
-
-                    u'Two playlists have this name':[
-                        u'c89078a6-0c35-4f53-88fe-21afdc51a414',
-                        u'86c69009-ea5b-4474-bd2e-c0fe34ff5484'
-                    ]
-                }
-            }
-
-        There is currently no support for retrieving automatically-created instant mixes
-        (see issue `#67 <https://github.com/simon-weber/Unofficial-Google-Music-API/issues/67>`__).
-
-        """
-
-        playlists = {}
-
-        if auto:
-            #playlists['auto'] = self._get_auto_playlists()
-            playlists['auto'] = {}
-        if user:
-            res = self._make_call(webclient.GetPlaylistSongs, 'all')
-            playlists['user'] = self._playlist_list_to_dict(res['playlists'])
-
-        return playlists
-
-    def _playlist_list_to_dict(self, pl_list):
-        ret = {}
-
-        for name, pid in ((p["title"], p["playlistId"]) for p in pl_list):
-            if not name in ret:
-                ret[name] = []
-            ret[name].append(pid)
-
-        return ret
-
-    def _get_auto_playlists(self):
-        """For auto playlists, returns a dictionary which maps autoplaylist name to id."""
-
-        #Auto playlist ids are hardcoded in the wc javascript.
-        #When testing, an incorrect name here will be caught.
-        return {u'Thumbs up': u'auto-playlist-thumbs-up',
-                u'Last added': u'auto-playlist-recent',
-                u'Free and purchased': u'auto-playlist-promo'}
-
     @utils.accept_singleton(basestring, 2)
     @utils.enforce_ids_param(2)
     @utils.enforce_id_param
     @utils.empty_arg_shortcircuit(position=2)
+    @utils.deprecated('prefer Mobileclient.add_songs_to_playlist')
     def add_songs_to_playlist(self, playlist_id, song_ids):
-        """Appends songs to a playlist.
+        """**Deprecated**: prefer :func:`Mobileclient.add_songs_to_playlist`.
+
+        Appends songs to a playlist.
         Returns a list of (song id, playlistEntryId) tuples that were added.
 
         :param playlist_id: id of the playlist to add to.
@@ -425,8 +321,11 @@ class Webclient(_Base):
     @utils.enforce_ids_param(2)
     @utils.enforce_id_param
     @utils.empty_arg_shortcircuit(position=2)
+    @utils.deprecated('prefer Mobileclient.remove_entries_from_playlist')
     def remove_songs_from_playlist(self, playlist_id, sids_to_match):
-        """Removes all copies of the given song ids from a playlist.
+        """**Deprecated**: prefer :func:`Mobileclient.remove_entries_from_playlist`.
+
+        Removes all copies of the given song ids from a playlist.
         Returns a list of removed (sid, eid) pairs.
 
         :param playlist_id: id of the playlist to remove songs from.